+++ src/main/java/net/frozenblock/lib/sound/mixin/LivingEntityMixin.java
@@ -18,7 +18,7 @@
 package net.frozenblock.lib.sound.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
-import net.frozenblock.lib.sound.api.damagesource.PlayerDamageSourceSounds;
+import net.frozenblock.lib.sound.api.damagesource.PlayerDamageTypeSounds;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.LivingEntity;
@@ -30,8 +30,8 @@ public abstract class LivingEntityMixin {
 
 	@ModifyReturnValue(method = "getHurtSound", at = @At("RETURN"))
 	private SoundEvent playHurtSound(SoundEvent original, DamageSource source) {
-		if (PlayerDamageSourceSounds.containsSource(source)) {
-			return PlayerDamageSourceSounds.getDamageSound(source);
+		if (PlayerDamageTypeSounds.containsSource(source.type())) {
+			return PlayerDamageTypeSounds.getDamageSound(source.type());
 		}
 		return original;
 	}